% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{buildChromSizes}
\alias{buildChromSizes}
\title{Build chromosome sizes object}
\usage{
buildChromSizes(assembly)
}
\arguments{
\item{assembly}{string, reference assembly identifier to chromosome sizes for}
}
\value{
named int, sizes of chromosome with respective names
}
\description{
Build chromosome sizes object
}
\examples{
\dontrun{
buildChromSizes("hg19")
}
}
