% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{buildGeneModels}
\alias{buildGeneModels}
\title{Build gene models}
\usage{
buildGeneModels(refAssembly)
}
\arguments{
\item{refAssembly}{string, reference assembly identifier to build gene
model for}
}
\value{
a list of four GRanges objects: genesGR, exonsGR, threeUTRGR,
fiveUTRGR
}
\description{
Build gene models
}
\examples{
\dontrun{
hg19models = buildGeneModels("hg19")
}
}
