% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{buildOpenSignalMatrix}
\alias{buildOpenSignalMatrix}
\title{Build open signal matrix}
\usage{
buildOpenSignalMatrix(assembly)
}
\arguments{
\item{assembly}{string, reference assembly identifier to build gene
model for}
}
\value{
data.frame, rows represent whole selection of open
chromatin regions across all cell types defined by ENCODE, columns are
individual cell types and values are normalized open chromatin signal values
}
\description{
Build open signal matrix
}
\examples{
\dontrun{
buildOpenSignalMatrix("hg19")
}
}
