% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{buildTSS}
\alias{buildTSS}
\title{Build TSS}
\usage{
buildTSS(assembly)
}
\arguments{
\item{assembly}{string, reference assembly identifier to TSS for}
}
\value{
GRanges object that consists of transcription start sites locations
}
\description{
Build TSS
}
\examples{
\dontrun{
tss = buildTSS("hg19")
}
}
