% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromSizes_hg19}
\alias{chromSizes_hg19}
\title{hg19 chromosome sizes}
\format{
A named vectors of lengths with one item per chromosome
}
\source{
BSgenome.Hsapiens.UCSC.hg19 package
}
\usage{
chromSizes_hg19(metadata=FALSE)
}
\arguments{
\item{metadata}{\code{logical} value indicating whether only
metadata should be returned or if the resource
should be loaded}
}
\value{
A vector of lengths
}
\description{
A dataset containing chromosome sizes for Homo sapiens hg19 genome assembly
}
\examples{
hg19c = chromSizes_hg19()

}
\keyword{datasets}
