% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromSizes_mm10}
\alias{chromSizes_mm10}
\title{mm10 chromosome sizes}
\format{
A named vectors of lengths with one item per chromosome
}
\source{
BSgenome.MMusculus.UCSC.mm10 package
}
\usage{
data(chromSizes_mm10)
}
\description{
A dataset containing chromosome sizes for Mus Musculus mm10 genome assembly
}
\keyword{datasets}
