% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromSizes_mm9}
\alias{chromSizes_mm9}
\title{mm9 chromosome sizes}
\format{
A named vectors of lengths with one item per chromosome
}
\source{
BSgenome.MMusculus.UCSC.mm9 package
}
\usage{
chromSizes_mm9(metadata=FALSE)
}
\arguments{
\item{metadata}{\code{logical} value indicating whether only
metadata should be returned or if the resource
should be loaded}
}
\value{
A vector of lengths
}
\description{
A dataset containing chromosome sizes for Mus musculus mm9 genome assembly
}
\examples{
mm9c = chromSizes_mm9()

}
\keyword{datasets}
