% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneModels_hg19}
\alias{geneModels_hg19}
\title{hg19 gene models}
\format{
A list of two GRanges objects, with genes and exons locations
}
\source{
EnsDb.Hsapiens.v75 package
}
\usage{
data(geneModels_hg19)
}
\description{
A dataset containing gene models for Homo Sapiens hg19 genome assembly
}
\keyword{datasets}
