% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneModels_hg38}
\alias{geneModels_hg38}
\title{hg38 gene models}
\format{
A list of two GRanges objects, with genes and exons locations
}
\source{
EnsDb.Hsapiens.v86 package
}
\usage{
geneModels_hg38(metadata=FALSE)
}
\arguments{
\item{metadata}{\code{logical} value indicating whether only
metadata should be returned or if the resource
should be loaded}
}
\value{
A list with two GRanges objects.
}
\description{
A dataset containing gene models for Homo sapiens hg38 genome assembly
}
\examples{
hg38GeneModels = geneModels_hg38()

}
\keyword{datasets}
