% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneModels_mm10}
\alias{geneModels_mm10}
\title{mm10 gene models}
\format{
A list of two GRanges objects, with genes and exons locations
}
\source{
EnsDb.Mmusculus.v79 package
}
\usage{
data(geneModels_mm10)
}
\description{
A dataset containing gene models for Mus Musculus mm10 genome assembly
}
\keyword{datasets}
