% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneModels_mm9}
\alias{geneModels_mm9}
\title{mm9 gene models}
\format{
A list of two GRanges objects, with genes and exons locations
}
\source{
TxDb.Mmusculus.UCSC.mm9.knownGene package
}
\usage{
geneModels_mm9(metadata=FALSE)
}
\arguments{
\item{metadata}{\code{logical} value indicating whether only
metadata should be returned or if the resource
should be loaded}
}
\value{
A list with two GRanges objects.
}
\description{
A dataset containing gene models for Mus musculus mm9 genome assembly
}
\examples{
mm9GeneModels = geneModels_mm9()

}
\keyword{datasets}
