% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicDistributionsData.R
\name{GenomicDistributionsData}
\alias{GenomicDistributionsData}
\title{GenomicDIstributionsData}
\source{
UCSC hg19, hg38, mm9, mm10 genome packages and ensembldb
}
\description{
Data (hg19, hg38, mm9, mm10) for the GenomicDistributions package.
}
\details{
The vignette details how to
access the data from ExperimentHub:
browseVignettes("GenomicDistributionsData")
Details on how the data files were created can be
found under the R directory in utils.R and build.R
}
\examples{
\dontrun{ 
library(ExperimentHub)
hub = ExperimentHub()
q = query(hub, "GenomicDistributionsData")
q[[1]]
q[["EH3472"]]

}
}
