% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loadTxDb}
\alias{loadTxDb}
\title{Load selected TxDb library}
\usage{
loadTxDb(genomeBuild)
}
\arguments{
\item{genomeBuild}{string, genome identifier}
}
\value{
loaded library
}
\description{
Load selected TxDb library
}
\examples{
\dontrun{
loadTxDb("hg19")
}
}
