% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TSS_mm10}
\alias{TSS_mm10}
\title{mm10 TSS locations}
\format{
A named vectors of lengths with one item per chromosome
}
\source{
EnsDb.Mmusculus.v79 package
}
\usage{
data(TSS_mm10)
}
\description{
A dataset containing chromosome sizes for Mus Musculus mm10 genome assembly
}
\keyword{datasets}
