% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TSS_mm9}
\alias{TSS_mm9}
\title{mm9 TSS locations}
\format{
A named vectors of lengths with one item per chromosome
}
\source{
TxDb.Mmusculus.UCSC.mm9.knownGene package
}
\usage{
TSS_mm9(metadata=FALSE)
}
\arguments{
\item{metadata}{\code{logical} value indicating whether only
metadata should be returned or if the resource
should be loaded}
}
\value{
A GRanges object with mm9 TSS locations.
}
\description{
A dataset containing Transcription Start Sites for the Mus musculus
mm9 genome assembly
}
\examples{
mm9TSS = TSS_mm9()

}
\keyword{datasets}
