% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chromSizes_mm9}
\alias{chromSizes_mm9}
\title{mm9 chromosome sizes}
\format{
A named vectors of lengths with one item per chromosome
}
\source{
BSgenome.MMusculus.UCSC.mm9 package
}
\usage{
data(chromSizes_mm9)
}
\description{
A dataset containing chromosome sizes for Mus Musculus mm9 genome assembly
}
\keyword{datasets}
