% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneModels_hg19}
\alias{geneModels_hg19}
\title{hg19 gene models}
\format{
A list of two GRanges objects, with genes and exons locations
}
\source{
EnsDb.Hsapiens.v75 package
}
\usage{
geneModels_hg19(metadata=FALSE)
}
\arguments{
\item{metadata}{\code{logical} value indicating whether only
metadata should be returned or if the resource
should be loaded}
}
\value{
A list with two GRanges objects.
}
\description{
A dataset containing gene models for Homo sapiens hg19 genome assembly
}
\examples{
hg19GeneModels = geneModels_hg19()

}
\keyword{datasets}
