% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneModels_hg38}
\alias{geneModels_hg38}
\title{hg38 gene models}
\format{
A list of two GRanges objects, with genes and exons locations
}
\source{
EnsDb.Hsapiens.v86 package
}
\usage{
data(geneModels_hg38)
}
\description{
A dataset containing gene models for Homo Sapiens hg38 genome assembly
}
\keyword{datasets}
