% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneModels_mm10}
\alias{geneModels_mm10}
\title{mm10 gene models}
\format{
A list of two GRanges objects, with genes and exons locations
}
\source{
EnsDb.Mmusculus.v79 package
}
\usage{
geneModels_mm10(metadata=FALSE)
}
\arguments{
\item{metadata}{\code{logical} value indicating whether only
metadata should be returned or if the resource
should be loaded}
}
\value{
A list with two GRanges objects.
}
\description{
A dataset containing gene models for Mus musculus mm10 genome assembly
}
\examples{
mm10GeneModels = geneModels_mm10()

}
\keyword{datasets}
