% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneModels_mm9}
\alias{geneModels_mm9}
\title{mm9 gene models}
\format{
A list of two GRanges objects, with genes and exons locations
}
\source{
TxDb.Mmusculus.UCSC.mm9.knownGene package
}
\usage{
data(geneModels_mm9)
}
\description{
A dataset containing gene models for Mus Musculus mm9 genome assembly
}
\keyword{datasets}
