% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TSS_mm10}
\alias{TSS_mm10}
\title{mm10 TSS locations}
\format{
A named vectors of lengths with one item per chromosome
}
\source{
EnsDb.Mmusculus.v79 package
}
\usage{
TSS_mm10(metadata=FALSE)
}
\arguments{
\item{metadata}{\code{logical} value indicating whether only
metadata should be returned or if the resource
should be loaded}
}
\value{
A GRanges object with mm10 TSS locations.
}
\description{
A dataset containing Transcription Start Sites for the Mus musculus
mm10 genome assembly
}
\examples{
mm10TSS = TSS_mm10()

}
\keyword{datasets}
