% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TSS_mm9}
\alias{TSS_mm9}
\title{mm9 TSS locations}
\format{
A named vectors of lengths with one item per chromosome
}
\source{
TxDb.Mmusculus.UCSC.mm9.knownGene package
}
\usage{
data(TSS_mm9)
}
\description{
A dataset containing chromosome sizes for Mus Musculus mm9 genome assembly
}
\keyword{datasets}
